Fix reflection metadata for ELF.
On ELF platforms we currently output an ELF note but with totally invalid
contents.  We should emit a proper ELF note.

At the same time, take the opportunity to change the MetadataSections structure
to use relative pointers for ELF, which then means that we can directly look
up the metadata locations by locating the note and reading it, without having
to worry about relocations.

On Windows we can also simplify things a little as we should at least be
able to emit a constant for the MetadataSections structure.

rdar://101749382