Improve path lookup.
On Windows, swift_getRootPath() should return the root path (i.e. outside
the "bin" directory).

On Linux, the path for the runtime looks like
<swift-root>/lib/swift/linux/libswiftCore.so
rather than <swift-root>/lib/swift/libswiftCore.so as you might expect.

On macOS, the path for the runtime *can* have "macosx" in it if we're using the
library from the build path.  That's not how it gets installed when it's part of
the operating system, but it's how we build it.

Also make swift_getAuxiliaryExecutablePath() actually check that the executable
exists.  This does mean it can return nullptr now, but it also means that it
can search through the various locations it supports to try to find the file
you're after, which is more robust.

rdar://103071801