Don't directly link CoreSymbolication, plus some other things.
Instead of trying to link to CoreSymbolication, dynamically load/link it.
Also tweak the API a bit more to add support for counted frame omission
(the idea being that if you hit a really huge recursion, we can capture
the interesting information without bothering with every single frame,
and indicate that we've done so).