Make sure we close all fds and execute swift-backtrace as the real UID.
We don't want to hand privileges to the swift-backtrace program.  It should
run as the user that executed it, *not* as any setuid user, and it should
also not have anything other than stdin/stdout/stderr open when it starts
(this code is Darwin specific; we'll need something similar on Linux).