[Darwin] Security improvement for the backtracer.
Make the backtracer drop privileges from the port it is passed (so that
it has a task read port, not a full task port).  As part of that change,
the crash handler has to suspend the process's threads itself, and
while it's doing that I've added some locks to stop crash races where
multiple threads are going to crash and try to enter the handler
simultaneously.