Polishing.
Updated the documentation.

Renamed "symbolicate" to "demangle" to reflect what it actually does.

Renamed the environment variable to SWIFT_BACKTRACE because that's easier
to type.

Changed the detection for the `$main` function to make it more robust.

Updated all the warnings emitted by the code in the runtime to start with
"swift runtime" so their source can easily be identified.